Refs 11838. Check if the topic is in the same domain
We can have more than one topic with the same name, given they are on
different participants.

In the backed data model, topics do not belong to the participant,
but the domain, so that two compatible topics in two different
participants are treated as the same topic.

Two topics are compatible if they have the same name and data type.
Therefore, to check if the topic of a discovered endpoint already
exists, we need to check:
 - If there is already topic with the same name
 - If that topic is in the current domain
 - If that topic uses the same data type
If not, it is a new topic

Signed-off-by: Iker Luengo <ikerluengo@eprosima.com>