New type for the entity queue
Instead of taking a complete entity, the queue will take the info needed
to create the entity and all the companion entities it needs. Therefore,
the queue template type must contain all this information.

This commit modifies the type, but does not change the queue yet

Instead of having a separate type for each entity kind (participant,
datareader, datawriter), we had to create a single kind with all the
attributes, even though some will not be used depending on the entity
kind (which had to be also added to the struct to know which values have
meaning and which are empty placeholders)

The reasons for this are:
 - This type is created on the DDS listener thread, from a
`xxxDiscoveryInfo` instance.
 - We cannot keep a reference to this instance,
because the queue is processing it in
another thread, and the original instance will be destroyed in the
listener thread.
 - We cannot copy the instance because the object is not copyable (at
least in Windows, the copy/constructor methods are not exported)
 - We cannot create an abstract hierarchy with a generic base class and
specializations for each entity kind,
and keeping the queue as a queue of the base class,
because the `push` method (which is generic in the
template) is the one doing the local copy, and would only copy the base
class attributes, therefore, losing all the specializations at the time
we are processing the element.
 - We cannot make a queue for each of the specializations because that
would only make the synchronization problem worse, as each queue will be
processed in its own thread
 - Probably we could overload the `push` method on the entity queue to
make it handle the copies correctly, but that will require another bunch
of auxiliary functions / classes and/or dynamic allocations. Since the
struct is samll enough, we didn't find it worth it, it would complicate
the design and make it more error prone with not much benefit.

Signed-off-by: Iker Luengo <ikerluengo@eprosima.com>