Move entity creation from listener to queue
Take the type created on the previous commit and adapt the queue for it:

- The DDS listener no longer creates the entity. It extracts the info
from the xxxDiscoveryInfo to build the EntityDiscoveryInfo, and calls
the queue
- The queue does the processing the listener was doing before, checking
the consistency of the entity to create, and creating it.
- The queue only handles PARTICIPANT, DATAREADER, DATAWRITER entities.
These are the only ones that are discovered by the listener. All the
other entities are created elsewhere directly on the database
(by traffic, or by endpoint discovery)
- Make sure that the user listener is called whenever an entity (or
sub-entity) is created/updated

Signed-off-by: Iker Luengo <ikerluengo@eprosima.com>