The listener no longer throws
The listener does not check the preconditions of the entities, so it
does no longer throw.

The queue itself will not throw either. In case of failure of
preconditions, the internal exception is catched in the `process_sample`
and an error log is issues, to avoid crashing the queue thread.

Signed-off-by: Iker Luengo <ikerluengo@eprosima.com>