Queue refactor (#114)
* New type for the entity queue

Instead of taking a complete entity, the queue will take the info needed
to create the entity and all the companion entities it needs. Therefore,
the queue template type must contain all this information.

This commit modifies the type, but does not change the queue yet

Instead of having a separate type for each entity kind (participant,
datareader, datawriter), we had to create a single kind with all the
attributes, even though some will not be used depending on the entity
kind (which had to be also added to the struct to know which values have
meaning and which are empty placeholders)

The reasons for this are:
 - This type is created on the DDS listener thread, from a
`xxxDiscoveryInfo` instance.
 - We cannot keep a reference to this instance,
because the queue is processing it in
another thread, and the original instance will be destroyed in the
listener thread.
 - We cannot copy the instance because the object is not copyable (at
least in Windows, the copy/constructor methods are not exported)
 - We cannot create an abstract hierarchy with a generic base class and
specializations for each entity kind,
and keeping the queue as a queue of the base class,
because the `push` method (which is generic in the
template) is the one doing the local copy, and would only copy the base
class attributes, therefore, losing all the specializations at the time
we are processing the element.
 - We cannot make a queue for each of the specializations because that
would only make the synchronization problem worse, as each queue will be
processed in its own thread
 - Probably we could overload the `push` method on the entity queue to
make it handle the copies correctly, but that will require another bunch
of auxiliary functions / classes and/or dynamic allocations. Since the
struct is samll enough, we didn't find it worth it, it would complicate
the design and make it more error prone with not much benefit.

Signed-off-by: Iker Luengo <ikerluengo@eprosima.com>

* Move entity creation from listener to queue

Take the type created on the previous commit and adapt the queue for it:

- The DDS listener no longer creates the entity. It extracts the info
from the xxxDiscoveryInfo to build the EntityDiscoveryInfo, and calls
the queue
- The queue does the processing the listener was doing before, checking
the consistency of the entity to create, and creating it.
- The queue only handles PARTICIPANT, DATAREADER, DATAWRITER entities.
These are the only ones that are discovered by the listener. All the
other entities are created elsewhere directly on the database
(by traffic, or by endpoint discovery)
- Make sure that the user listener is called whenever an entity (or
sub-entity) is created/updated

Signed-off-by: Iker Luengo <ikerluengo@eprosima.com>

* fixup

Signed-off-by: Iker Luengo <ikerluengo@eprosima.com>

* Locator must be created before endpoint inserted

The queue is responsible of creating all the companion entities (this
reduces the spread of where the entities are created). Therefore, the
locator must be already created when the endpoint is inserted

This is coherent with the behavior with the Topic, for example

Signed-off-by: Iker Luengo <ikerluengo@eprosima.com>

* Adapt queue tests

- Remove all tests not related to PARTICIPANT, DATAREADER, DATAWRITER
- Some preconditions had to be modified to the new API
- Some new tests to reflect the fact that the locators and topics
are created on the queue
- Using strict mocks

Signed-off-by: Iker Luengo <ikerluengo@eprosima.com>

* The listener no longer throws

The listener does not check the preconditions of the entities, so it
does no longer throw.

The queue itself will not throw either. In case of failure of
preconditions, the internal exception is catched in the `process_sample`
and an error log is issues, to avoid crashing the queue thread.

Signed-off-by: Iker Luengo <ikerluengo@eprosima.com>

* Locators must be inserted before entities

Signed-off-by: Iker Luengo <ikerluengo@eprosima.com>

* Refs #12137. Fix tests after rebase

Signed-off-by: Ricardo González Moreno <ricardo@richiware.dev>

* Refs #12137. Fix failed tests on CI

Signed-off-by: Ricardo González Moreno <ricardo@richiware.dev>

* Fix coredump when unique ptr of consumer thread is reset

Signed-off-by: RaulSanchez <raul@eprosima.com>

Co-authored-by: Ricardo González Moreno <ricardo@richiware.dev>
Co-authored-by: RaulSanchez <raul@eprosima.com>